Added project Solution/Repository/ImportDataToDb
With this, one can import the Exin's full File Repository data into an
Exin database. This can be MS SQL or SQLite, which can be configured in
app.config

Also removed folder Solution/Repository/DB/FullScript/Archive. This was
added for reserving in the previous commit, and promised to be removed