Added project Solution/WEB
This was an ASP.NET MVC 4 based web UI for the app. University homework as
well. Wasn't ever meant for production; it's not stable and it's
functionality is not the most rich one :)
Originally, it was removed from the project (long) before it got to source
control. (In the there was a Windows 8 update which messed up with
System.Web.Mvc...) Now, after a little refactors, put it back into the
project, for preservation of course. It will be removed (with WinForms and
UnitTests) soon; moreover, this full commit will be rollback via a
cleanup.