Clenup: Removed Solution/Repository/FileRepoDeveloper
It contains local data.
Another folder named
	Solution/Repository/FileRepoDeveloper.template/
has been created instead, and will contain sandbox data for Debug
mode in the future. Creating a local copy of it is optional,
because the app will init an empty repo if does not find any. Only
that case, the DeveloperRepo will be empty.

TODO fill it with sandbox data

(Also updated helpers)