Localization: Unit, Category (part 1)
They had ID, Name and DisplayName properties; now the last one got to
displaynameS, and this one get localized, and contains from now on data
like this:
"en-US:Food;hu-HU:Kaja;"
The modified SQLite and MsSql scripts are done, as well as the xml parser.
Also tested it with (pure) AdoNet, yet only the ImportDataToDb works; but
that for both.
We can't read the data from db yet, because the edmx files and the db
specific entities are still to come.