BugFix: Regression ( [Serializable] ) (rollback)
The modified Category and Unit classes weren't marked as Serializable, so
the BinaryFormatter serializer failed. Replaced it with Json.NET
serialization; because we will need that later any case.
Realized the app does not work with it yet, so restored the original
binary formatter solution (and marked the above classes as Serializable)