BugFix: MsSql with AdoNet could not insert single expense
This bug could be there for some time now, it does not seem to be related to the refactor.
Inserting eg daily expenses looks like this:
	ClearDay
	(clear the query)
	InsertNewData
Here the "(clear the query)" part was wrong, only the CommandText got cleared, the Command.Paramters not.
Interesting that it only caused problem by the MsSql Provider, not for SQLite Provider.
The error was that the @Date property was duplicated. Interesting again that the other properties why were not duplicated...