BugFix: Inconsistent currency on UI
There is only HUF as currency in the app, though the UI's language can be set to English already. There were "$"-s appeared instead of "Ft" by some UI elements when the language was set to English.
Fixed them, though it's just a temporary fix, because later there have to be more currency in the system.