BugFix: MonthlyExpenses ListView: Duble click on header
That probably means 2 fast clicks in reality; but will be handled as
DoubleClick in the system of course. That's why the DoubleClick handler
also will be called, so if there is a SelectedItem in the ListView, the
app will jump to its DailyExpense item. Because if one double clicks an
expense row in the MonthlyExpense ListView, that will happen.
Fixed now