Selected-Edited item (1)
* That row, which is currently edited in eg "Add new daily expense", will be highlighted (bold)
* The Add button's label will be "Copy" if a Selected-Edited item exists.

This is the fist version, started to implement using indexes. This was a very bad idea; because indexes should be maintained while add and remove operations. Much easier way is to prefer SelectedItem instead of SelectedIndex.
Another mistake with this implementation, that logic is injected into Add and Remove button click handlers. It should be bound to the Actual..Item instead