Updated Helpers: SQLite vs Check Constraints
cat Helpers\TODO\MS SQL vs Category 0.txt

The problem is not in MS SQL, it does work properly!
There are Check Constraints on the TransactionItem table; one of them gets violated, that is visible on the picture.
The real problem is that the SQLite file does not have these constraints! The tool with which we convert the MS SQL db into SQLite does not convert the Check Constraints!
...