BugFix: Wrong localized validation message
Created LocalizedDisplayNameAttribute class to give localized display names for the entities

Previous, non-working solutions:
* System.ComponentModel.DisplayNameAttribute: Not localized
* System.ComponentModel.DataAnnotations.DisplayAttribute: Localized, but it
caches the values fetched from the ResourceManager. As long as runtime
language changing is not possible in the app, it should no cause problems; but
it seems it does. From a while, it only shows the English values.
** First thought was that it probably it prepares
	it's caches before the app's language settings are loaded from config.
	But English strings will be shown also when the OS language is Hunagarian